pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = [
  "setuptools",
]

[project]
name = "clx"
version = "0.0.1"
description = "Toolkit for running experiments and developing classifiers."
authors = [{ name = "Free Law Project", email = "info@free.law" }]
requires-python = ">=3.13"
readme = "README.md"
license = "BSD-2-Clause"
keywords = [
    "legal",
    "courts",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

urls.Repository = "https://github.com/freelawproject/classifier-experiments"
urls."Organisation Homepage" = "https://free.law/"

dependencies = [
  "boto3>=1.28.57",
  "click>=8.3.0",
  "datasets>=4.3.0",
  "django>=5.2.7",
  "django-postgres-copy>=2.8.0",
  "django-shortuuid>=0.1.2",
  "dspy>=3.0.3",
  "hf_transfer>=0.1.9",
  "litellm>=1.79.1",
  "lmdb>=1.7.5",
  "pandas>=2.3.3",
  "pgvector>=0.4.2",
  "psycopg2-binary>=2.9.11",
  "python-dotenv>=1.2.1",
  "requests>=2.32.5",
  "runpod>=1.8.1",
  "sentencepiece>=0.2.1",
  "scikit-learn>=1.7.2",
  "simplejson>=3.20.2",
  "tabulate>=0.9.0",
  "tensorboard>=2.20.0",
  "tqdm>=4.67.1",
  "transformers[torch]>=4.57.1",
]

[project.optional-dependencies]
dev = [
  "pre-commit>=4.3.0",
  "tox-uv>=1.29.0",
]

[project.scripts]
clx = "clx:cli.cli"

[tool.setuptools.packages.find]
include = ["clx*"]

[tool.ruff]
lint.isort.known-first-party = [
  "clx",
]
line-length = 79
lint.extend-safe-fixes = [
  # non-pep585-annotation
  "UP006",
]
lint.select = [
  # flake8-bugbear
  "B",
  # flake8-comprehensions
  "C4",
  # pycodestyle
  "E",
  # Pyflakes errors
  "F",
  # isort
  "I",
  # flake8-simplify
  "SIM",
  # flake8-tidy-imports
  "TID",
  # pyupgrade
  "UP",
  # Pyflakes warnings
  "W",
]
lint.ignore = [
  # flake8-bugbear opinionated rules
  "B9",
  # line-too-long
  "E501",
  # suppressible-exception
  "SIM105",
  # if-else-block-instead-of-if-exp
  "SIM108",
  # printf-string-formatting
  "UP031",
]

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "clx.settings"
